requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
executing==2.1.0
fastapi==0.115.5
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logfire==2.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mslex==1.3.0
narwhals==1.14.3
numpy==2.1.3
opentelemetry-api==1.28.2
opentelemetry-distro==0.49b2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-asyncio==0.49b2
opentelemetry-instrumentation-dbapi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-grpc==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
opentelemetry-instrumentation-jinja2==0.49b2
opentelemetry-instrumentation-logging==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-instrumentation-sqlite3==0.49b2
opentelemetry-instrumentation-starlette==0.49b2
opentelemetry-instrumentation-system-metrics==0.49b2
opentelemetry-instrumentation-threading==0.49b2
opentelemetry-instrumentation-tornado==0.49b2
opentelemetry-instrumentation-tortoiseorm==0.49b2
opentelemetry-instrumentation-urllib==0.49b2
opentelemetry-instrumentation-urllib3==0.49b2
opentelemetry-instrumentation-wsgi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
protobuf==5.29.0
psutil==6.1.0
pyarrow==18.1.0
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.40.2
taskipy==1.14.1
tenacity==9.0.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
wrapt==1.17.0
zipp==3.21.0
